fix: enhance code
