+++ src/components/TimelineMarker/index.js
@@ -1,124 +1,22 @@
-import React, { useRef, useContext, useEffect, useCallback, useMemo } from 'react';
+import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
-import RenderIf from '../RenderIf';
-import { useUniqueIdentifier } from '../../libs/hooks';
 import { ActivityTimelineContext } from '../ActivityTimeline/context';
 import CalendarIcon from './calendarIcon';
-import StyledLi from './styled/li';
-import StyledColumnLeft from './styled/columnLeft';
-import StyledIconContainer from './styled/iconContainer';
-import StyledContentContainer from './styled/contentContainer';
-import StyledHeader from './styled/header';
-import StyledLabel from './styled/label';
-import StyledDatetime from './styled/datetime';
-import StyledDescription from './styled/description';
-import StyledBody from './styled/body';
-import ExpandCollapseButton from './expandCollapseButton';
-import removeItemFromArray from '../AccordionSection/removeItemFromArray';
-import isInArray from '../AccordionSection/isInArray';
+import AccordionTimelineMarker from './accordionMarker';
+import BasicTimelineMarker from './basicMarker';
 
 /**
  * The TimelineMarker displays one event of an item's timeline. It's generally used to compose the ActivityTimeline component.
  * @category Layout
  */
 export default function TimelineMarker(props) {
-    const {
-        icon,
-        name,
-        isLoading,
-        label,
-        description,
-        datetime,
-        children,
-        className,
-        style,
-    } = props;
-    const {
-        isVariantAccordion,
-        activeNames,
-        multiple,
-        privateRegisterMarker,
-        privateUnregisterMarker,
-        privateOnToggleMarker,
-    } = useContext(ActivityTimelineContext);
-    const generatedName = useUniqueIdentifier('timeline-marker');
-    const currentName = useMemo(() => name || generatedName, [generatedName, name]);
-    const stepRef = useRef();
-
-    useEffect(() => {
-        privateRegisterMarker(stepRef.current, { currentName });
-
-        return () => {
-            privateUnregisterMarker(stepRef, currentName);
-        };
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
-    const resolveActiveNamesWhenMultiple = useCallback(() => {
-        if (activeNames === undefined) {
-            return [currentName];
-        }
-        if (isInArray(activeNames, currentName)) {
-            return removeItemFromArray(activeNames, currentName);
-        }
-        return activeNames.concat([currentName]);
-    }, [currentName, activeNames]);
-
-    const resolveActiveNames = useCallback(() => {
-        if (multiple) {
-            return resolveActiveNamesWhenMultiple();
-        }
-        if (currentName === activeNames) {
-            return '';
-        }
-        return currentName;
-    }, [currentName, multiple, activeNames, resolveActiveNamesWhenMultiple]);
-
-    const handleToggleMarker = useCallback(
-        event => privateOnToggleMarker(event, resolveActiveNames()),
-        [privateOnToggleMarker, resolveActiveNames],
-    );
-
-    const isExpanded = useMemo(() => {
-        if (isVariantAccordion) {
-            if (multiple && Array.isArray(activeNames)) {
-                return isInArray(activeNames, currentName);
-            }
-            return activeNames === currentName;
-        }
-        return true;
-    }, [isVariantAccordion, currentName, activeNames, multiple]);
-
-    const shouldRenderChildren = useMemo(() => isExpanded && !isLoading, [isLoading, isExpanded]);
-
-    return (
-        <StyledLi data-id="timeline-marker-li" className={className} style={style}>
-            <StyledColumnLeft>
-                <StyledIconContainer>
-                    <RenderIf isTrue={isVariantAccordion}>
-                        <ExpandCollapseButton
-                            isExpanded={isExpanded}
-                            isLoading={isLoading}
-                            onClick={handleToggleMarker}
-                        />
-                    </RenderIf>
-                    {icon}
-                </StyledIconContainer>
-            </StyledColumnLeft>
-            <StyledContentContainer>
-                <div>
-                    <StyledHeader>
-                        <StyledLabel>{label}</StyledLabel>
-                        <StyledDatetime>{datetime}</StyledDatetime>
-                    </StyledHeader>
-                    <StyledDescription>{description}</StyledDescription>
-                </div>
-                <RenderIf isTrue={shouldRenderChildren}>
-                    <StyledBody>{children}</StyledBody>
-                </RenderIf>
-            </StyledContentContainer>
-        </StyledLi>
-    );
+    const context = useContext(ActivityTimelineContext);
+    if (context) {
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        return <AccordionTimelineMarker {...props} />;
+    }
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    return <BasicTimelineMarker {...props} />;
 }
 
 TimelineMarker.propTypes = {
