+++ src/components/TimelineMarker/styled/columnLeft.js
@@ -2,8 +2,6 @@ import styled from 'styled-components';
 import attachThemeAttrs from '../../../styles/helpers/attachThemeAttrs';
 import { getContrastText, replaceAlpha } from '../../../styles/helpers/color';
 
-const getColor = props => replaceAlpha(getContrastText(props.palette.background.main), 0.1);
-
 const StyledColumnLeft = attachThemeAttrs(styled.div)`
     display: flex;
     justify-content: center;
@@ -12,7 +10,8 @@ const StyledColumnLeft = attachThemeAttrs(styled.div)`
 
     ::before {
         content: '';
-        background-color: ${props => getColor(props)};
+        background-color: ${props =>
+            replaceAlpha(getContrastText(props.palette.background.main), 0.1)};
         height: 100%;
         width: 1px;
         position: absolute;
